pyproject.toml
[project]
name = "transformer-ocr-llm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "editdistance>=0.8.1",
    "ipywidgets>=8.1.8",
    "jiwer>=4.0.0",
    "jupyter>=1.1.1",
    "numpy>=2.4.2",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.2.0",
]
